feat: pss-cln-def-fold-distribution-001 - Fold package asset assembly into distribution with list/resolve parity
Move packageassets and promptassets implementation into distribution-owned private modules, leave transitional public shims, and reconcile ownership and coverage manifests for the folded asset assembly surface.
