+++ pkg/services/factory_definitions/internal/fold_sibling_boundary_test.go
@@ -18,6 +18,8 @@ var foldedPublicSiblingSuffixes = []string{
 	"/loadedsource",
 	"/snapshotcapture",
 	"/packagedinstallation",
+	"/packages/packageassets",
+	"/packages/promptassets",
 	"/namedfactories",
 }
