+++ pkg/services/factory_definitions/internal/services/distribution/promptassets/testcomposition_test.go
@@ -1,0 +1,45 @@
+package promptassets_test
+
+import (
+	"github.com/portpowered/infinite-you/pkg/platform/directoryreplace"
+	platformfilesystem "github.com/portpowered/infinite-you/pkg/platform/filesystem"
+	"github.com/portpowered/infinite-you/pkg/platform/inboxgitkeep"
+
+	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
+	factorydefinitiontestcomposition "github.com/portpowered/infinite-you/pkg/services/factory_definitions/internal/testcomposition"
+	factorymapping "github.com/portpowered/infinite-you/pkg/transports/mapping/factoryconfig"
+	authoredmapping "github.com/portpowered/infinite-you/pkg/transports/mapping/factoryconfig/authored"
+	"github.com/portpowered/infinite-you/pkg/transports/mapping/validationentry"
+)
+
+var factorydefinitioncomposition = func() factorydefinitiontestcomposition.Composition {
+	mapper := factorymapping.NewFactoryConfigMapper()
+	fileSystem := platformfilesystem.Local{}
+	var composition factorydefinitiontestcomposition.Composition
+	composition = factorydefinitiontestcomposition.New(factorydefinitiontestcomposition.Representation{
+		DecodeFactory:     factorymapping.ExpandFactoryConfigForRuntimeLoad,
+		DecodeAuthored:    mapper.Expand,
+		EncodeFactory:     factorymapping.MarshalCanonicalFactoryConfig,
+		NormalizeAuthored: authoredmapping.AuthoredFactoryConfigForExpandedLayout,
+		ParseWorker:       authoredmapping.ParseWorkerConfig,
+		ParseWorkstation:  authoredmapping.ParseWorkstationConfig,
+		ParseBody:         authoredmapping.ParseAgentsBody,
+		RenderWorker:      authoredmapping.RenderWorkerAgentsMarkdown,
+		RenderWorkstation: authoredmapping.RenderWorkstationAgentsMarkdown,
+		RenderBody:        authoredmapping.RenderAgentsBody,
+		SafeLayoutSegment: authoredmapping.SafeFactoryLayoutSegment,
+		SafePromptPath:    authoredmapping.SafePromptFilePath,
+		MapPersistence: func(payload []byte) (factorydefinitions.DefinitionValidationRequest, error) {
+			return validationentry.MapFactoryJSONForPersistence(payload, composition.LoadCanonicalJSON)
+		},
+	}, fileSystem, directoryreplace.Local{}, factorydefinitiontestcomposition.Effects{
+		Loading:             fileSystem,
+		AuthoredReader:      fileSystem,
+		AuthoredWriter:      fileSystem,
+		Persistence:         fileSystem,
+		NamedPaths:          fileSystem,
+		NamedFactoryCatalog: fileSystem,
+		InboxSentinels:      inboxgitkeep.NewLocal(fileSystem),
+	})
+	return composition
+}()
